### 2022-07-05 #4549 @editorialbot
[PRE REVIEW]: WILL - Weighted Injector of Luminous Lighthouses
**Submitting author:** @josephwkania (<a href="http://orcid.org/0000-0002-3354-3859">Joseph Kania</a>)
**Repository:** https://github.com/josephwkania/will/
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.0
