### 2022-07-06 #4549 @dfm
I'm happy to edit! I'm traveling with limited internet access for the next week so my initial editorial pass and search for reviewers may be a little delayed, but I'll keep you posted.

@josephwkania — As @danielskatz [mentions above](https://github.com/openjournals/joss-reviews/issues/4549#issuecomment-1175524517), please do add a LICENSE as soon as possible.
