### 2022-07-14 #4549 @dfm
@josephwkania — I'm going through this submission now and I was hoping you could provide a few more details about the relationship between this submission and (a) the required dependency [jess](https://github.com/josephwkania/jess), and (b) the intellectual predecessor of `will.create`, [injectfrb](https://github.com/liamconnor/injectfrb). Neither of these other projects are open source licensed, and the latter isn't credited in the source. Thanks!
