### 2022-07-27 #4549 @josephwkania
I've updated `jess`'  licensee to GPL-3.

[scintillation](https://github.com/josephwkania/will/blob/5e1c7ca63c91902ccc45bffeaa8a9177bb026cb0/will/create.py#L504) is  based on
https://github.com/liamconnor/injectfrb/blob/a4dd5f22438ba7bdfaa2eb792eb54736cde53fed/injectfrb/simulate_frb.py#L77
which I modified to make the scintiles (what I consider) more realistic.

[scatter_profile](https://github.com/josephwkania/will/blob/5e1c7ca63c91902ccc45bffeaa8a9177bb026cb0/will/create.py#L534) is based on https://github.com/liamconnor/injectfrb/blob/a4dd5f22438ba7bdfaa2eb792eb54736cde53fed/injectfrb/simulate_frb.py#L111

I've added these references to the function's doc strings.
