### 2022-10-19 #4692 @liamconnor
I commend the author on a well-documented and sophisticated pulsar/FRB simulation package.

Okay, so I've installed WILL on a couple of different machines, gone through the tutorials, and everything seems to run okay. It's fairly easy to install, but I wouldn't call it "lightweight". Maybe that's fine. This is presumably due to its dependence on YOUR.

The statement of need in the paper is reasonably compelling, but I think it would benefit from some specific use cases. For example, will this code be used for real-time injections on CHIME or CHORD? The second author, Kevin Bandura, is of course a central figure in both. Or is WILL better for offline FRB simulations?

Maybe I'm missing it, but I don't see anything about how WILL handles intrachannel dispersion smearing and finite temporal sampling. I know that one of the motivations for writing "simpulse", which is cited as a single-pulse simulator in the WILL README, was that previous codes had not accounted for smearing in a physically realistic way. If pulses are not simulated from voltages, then ones needs a way to simulate a high time/freq resolution pulse being randomly sampled by an instrument with finite resolution in both of those dimensions.

Does WILL have functionality to inject real pulses / pulsars into data? CHIME/FRB has thousands of bursts that could in principle be injected directly into data to test other software pipelines, and completeness to different burst morphologies.
