+++ will/create.py
@@ -42,12 +42,12 @@ def log_normal_from_stats(median: float, std: float, size: int) -> np.ndarray:
     and standard deviation.
 
     Args:
-        median - median of resulting distrabution
+        median - median of resulting distribution
 
         std - Standard deviation of returned samples
 
     Returns:
-        `size` numbers from the sampled from lognormal distrabution
+        `size` numbers from the sampled from lognormal distribution
     """
     mu = np.log(median)
     sigma_guess = np.sqrt(mu - 0.5 * np.log(std**2))
@@ -84,7 +84,7 @@ def quicksort(
         sort_fraction - The fraction of the array to
                         stop sorting.
 
-        sort_assend - Sort increating
+        sort_assend - Sort increasing
 
     Returns:
         None - Sort in place
@@ -191,7 +191,7 @@ def skewed_gauss(
 
         x_sig - Horizontal sigma
 
-        y_sig - Verticel Sigma
+        y_sig - Vertical Sigma
 
         theta - Rotation angle increasing counterclockwise [radians]
 
@@ -225,7 +225,7 @@ def pulse_with_tail(times: np.ndarray, tau: float = 50) -> np.ndarray:
     Args:
         Times - Time array
 
-        tau - With paramater
+        tau - With parameter
 
     Returns:
         pulse profile
@@ -326,7 +326,7 @@ class pulse_with_tail_dist(stats.rv_continuous):
             norm_constant - Normalization Constant
 
         Returns:
-            Values sampled from pulse with tail distrution
+            Values sampled from pulse with tail distribution
         """
         # center: int
         return (1.0 / norm_const) * pulse_with_tail(times, tau=tau)
@@ -342,7 +342,7 @@ def gauss_with_tail_cdf(times: np.ndarray, tau: float) -> np.ndarray:
         tau - Pulse width
 
     Returns:
-        Values sampled from pulse with tail distrution
+        Values sampled from pulse with tail distribution
 
     Notes:
         based on
@@ -372,13 +372,16 @@ def arbitrary_array_cdf(
 
 
     Returns:
-        Values sampled from pulse with tail distrution
+        Values sampled from pulse with tail distribution
 
     Notes:
         based on
         https://harry45.github.io/blog/2016/10/Sampling-From-Any-Distribution
     """
     logging.debug("Sampling given array")
+    # if array has negatives, pulse_cdf will not be monotonic
+    # and interpolation is difficult
+    assert array.min() >= 0, "Probability array must be non-negative!"
     pulse_cdf = np.cumsum(array)
     pulse_cdf /= np.max(pulse_cdf)
     cdf_interp = interpolate.interp1d(pulse_cdf, locations)
@@ -660,6 +663,8 @@ class SimpleGaussPulse:
 
         center_freq - Center Frequency in MHz
 
+        chan_freq - Array of channel frequencies in MHz
+
         tsamp - sampling time of dynamic spectra in second
 
         spectra_index_alpha - spectral index around center_freq
@@ -797,6 +802,84 @@ class SimpleGaussPulse:
 
 
 @dataclass
+class TwoDimensionalPulse:
+    """
+    Create a pulse from a 2D pulse Probability Distribution Function (PDF).
+
+    Args:
+        pulse_pdf - The 2D array containing pulse the pulse profile at
+                    0 DM.
+
+        chan_freq - Array of channel frequencies in MHz
+
+        tsamp - sampling time of dynamic spectra in second
+
+        dm - Dispersion Measure
+
+    """
+
+    pulse_pdf: np.ndarray
+    chan_freqs: np.ndarray
+    tsamp: float
+    dm: float
+
[…]
