### 2022-11-02 #4692 @abatten
I also commend the author. This is quite a sophisticated simulation package. I'm not personally as familiar with the needs and requirements for injection pipelines, so I'll leave those recommendations to @liamconnor.

But the software seems to work with the examples provided. And the main three functions are documented. Although I would add a note somewhere stating which parameters are required and which are optional (if any). This would make it clear what the minimum requirements are to get `will` to do *something*.

I was able to install the package and run almost everything. However, I did encounter a couple of issues I found when using the package. Nothing too major:
* When running the optimise.ipynb, the package `cupy` is required but isn't downloaded with any requirements file.
* Running `pytest` in the `will` directory will run the tests. However, 9 of the tests failed. This was run in a fresh conda environment in python 3.10. All of them were `unsupported operand type(s) for *: 'NoneType' and 'NoneType'`.
```
=============================================================== short test summary info ================================================================
FAILED tests/test_detect.py::test_find_first_pulse - TypeError: unsupported operand type(s) for *: 'NoneType' and 'NoneType'
FAILED tests/test_detect.py::test_location_pules - TypeError: unsupported operand type(s) for *: 'NoneType' and 'NoneType'
FAILED tests/test_detect.py::test_search_file - TypeError: unsupported operand type(s) for *: 'NoneType' and 'NoneType'
FAILED tests/test_detect.py::test_extract_pulses - TypeError: unsupported operand type(s) for *: 'NoneType' and 'NoneType'
FAILED tests/test_inject.py::TestInjectConstant::test_inject_constant - TypeError: unsupported operand type(s) for *: 'NoneType' and 'NoneType'
FAILED tests/test_inject.py::TestInjectConstant::test_chan_error - TypeError: unsupported operand type(s) for *: 'NoneType' and 'NoneType'
FAILED tests/test_inject.py::TestInjectConstant::test_gulp_error - TypeError: unsupported operand type(s) for *: 'NoneType' and 'NoneType'
FAILED tests/test_inject.py::TestInjectDistro::test_inject_distro - TypeError: unsupported operand type(s) for *: 'NoneType' and 'NoneType'
FAILED tests/test_inject.py::TestInjectDistro::test_gulp_error - TypeError: unsupported operand type(s) for *: 'NoneType' and 'NoneType'
============================================================= 9 failed, 39 passed in 5.69s =============================================================
```
