### 2022-11-04 8708fae
Added multi 2d convolution and channel dm smearing functions w/ tests, moved some functions from create to calculate
