+++ tests/test_calculate.py
@@ -1,0 +1,250 @@
+#!/usr/bin/env python3
+"""
+Test will.calculate
+"""
+import numpy as np
+
+from will import calculate
+
+rng = np.random.default_rng(2022)
+
+
+def test_std_min_func():
+    """
+    Test the std minimization function
+    We know where it should give zeros
+    """
+    # if sigma is zero, std must also be zero
+    assert calculate.std_min_func(0, 1, 0) == 0
+    # calculate the expected variance for mu=sigma=1
+    assert calculate.std_min_func(1, 1, np.sqrt((np.exp(1) - 1) * np.exp(2 + 1))) == 0
+
+
+def test_log_normal_from_stats():
+    """
+    Test if the properties of the log normal distro are
+    expected
+    """
+    median = 5e5
+    std = 0.1 * median
+    num_samples = 100
+    distro = calculate.log_normal_from_stats(median, std, num_samples)
+
+    # 17% seems generous enough
+    assert np.std(distro) / std - 1 < 0.17
+    assert np.median(distro) / median - 1 < 0.17
+    assert len(distro) == num_samples
+
+
+class TestQuickSort:
+    """
+    Test the quicksort
+    """
+
+    def setup_class(self):
+        """
+        Random array to sort
+        """
+        self.rands = rng.normal(size=20)
+
+    def test_sort_assent(self):
+        """
+        Test quicksort assending
+        """
+        rands_copy = self.rands.copy()
+        calculate.quicksort(rands_copy, sort_assend=True)
+        np.testing.assert_allclose(rands_copy, np.sort(self.rands))
+
+    def test_sort_desend(self):
+        """
+        Test quicksort desending
+        """
+        rands_copy = self.rands.copy()
+        calculate.quicksort(rands_copy, sort_assend=False)
+        np.testing.assert_allclose(rands_copy, np.sort(self.rands)[::-1])
+
+    def test_subsort_array(self):
+        """
+        Split an array in two and split
+        """
+        split = len(self.rands) // 2
+        array_split = np.concatenate(
+            [np.sort(self.rands[:split]), np.sort(self.rands[split:])[::-1]]
+        )
+        quicksort_split = self.rands.copy()
+        quicksort_split = calculate.sort_subarrays(quicksort_split, num_subarrays=2)
+        np.testing.assert_array_equal(quicksort_split, array_split)
+
+
+class TestCalculateDMWidths:
+    """
+    Test the DM width calculator.
+    """
+
+    def setup_class(self):
+        """
+        Set up constants, use GREENBURST numbers
+        """
+        self.nchans = 4096
+        self.chan_freqs = np.linspace(1919.8828125, 960.1171875, 4096)
+        self.chan_width = 0.234375
+
+    def test_zero_dm(self):
+        """
+        Should return zeros
+        """
+        assert not np.any(
+            calculate.calculate_dm_widths(0, self.chan_width, self.chan_freqs)
+        )
+
+    def test_ends(self):
+        """
+        Test the end values, lengths, and all positive.
+        """
+        widths = calculate.calculate_dm_widths(100, self.chan_width, self.chan_freqs)
+        assert self.chan_freqs.shape == widths.shape
+        assert widths[0] == calculate.calculate_dm_widths(
+            100, self.chan_width, self.chan_freqs[0]
+        )
+        assert widths[-1] == calculate.calculate_dm_widths(
+            100, self.chan_width, self.chan_freqs[-1]
+        )
+        assert np.all(widths > 0)
+
+
+class TestCalculateDMBoxcarWidths:
+    """
+    Test the DM width calculator.
+    """
+
+    def setup_class(self):
+        """
[…]

[…]
