### 2022-11-04 82ef84c
Added interchannel dm smearing option to GaussPulse, fixed spelling
