### 2022-11-04 55db71a
Updated README to reflect calculate addition; added presto to pulsar synth list
