### 2022-11-11 ec5e41c
Added noise calculators and helper funcs to calculator with tests
