+++ tests/test_calculate.py
@@ -2,12 +2,20 @@
 """
 Test will.calculate
 """
-# Can't use inits with pytest, this error is unavoidable
-# pylint: disable=W0201
+from unittest import mock
+
 import numpy as np
 import pytest
+from your.formats.filwriter import make_sigproc_object
 
-from will import calculate
+from will import calculate, create
+
+# Can't use inits with pytest, this error is unavoidable
+# pylint: disable=W0201
+
+# pylint: disable=redefined-outer-name
+# The pytest.fixture needs to be redefined
+
 
 rng = np.random.default_rng(2022)
 
@@ -244,7 +252,6 @@ class TestConvolveMultiBoxcar:
         """
         If profile is above two, raise NotImplementedError
         """
-        print(self.profile[:, None].shape)
         with pytest.raises(NotImplementedError):
             calculate.convolve_multi_boxcar(
                 self.profile[:, None, None], self.sqrt_array
@@ -264,3 +271,212 @@ def test_boxcar_convolved():
     assert convolved.argmax() == 1
     assert convolved[0] < convolved.max()
     assert convolved[2] < convolved.max()
+
+
+class TestMedianDetrend:
+    """
+    Test the median_line_detrend.
+    """
+
+    @staticmethod
+    def test_base():
+        """
+        Test no padding
+        """
+        num_samps = 2**14
+        noise = rng.normal(size=num_samps)
+        times = np.linspace(0, 1, num_samps)
+        noise_with_tend = noise + 5 * times + 4
+        detrend = calculate.median_line_detrend(noise_with_tend, num_sample=1024)
+        assert detrend.size == num_samps
+        fit = np.polyfit(times, detrend, deg=1)
+        assert (np.abs(fit) < np.array((0.1, 0.1))).all()
+
+    @staticmethod
+    def test_pad():
+        """
+        Test padding.
+        """
+        num_samps = 2**14 - 1
+        noise = rng.normal(size=num_samps)
+        times = np.linspace(0, 1, num_samps)
+        noise_with_tend = noise + 5 * times + 4
+        detrend = calculate.median_line_detrend(noise_with_tend, num_sample=1024)
+        assert detrend.size == num_samps
+        fit = np.polyfit(times, detrend, deg=1)
+        assert (np.abs(fit) < np.array((0.1, 0.1))).all()
+
+
+def test_calculate_noises_multi():
+    """
+    Make a Gaussian time series, check if
+    it integrates down as expected.
+    """
+
+    time_series = rng.normal(size=2**14)
+    boxcar_lengths = np.array([2**x for x in range(0, 8)])
+    boxcar_array, max_boxcar = calculate.generate_boxcar_array(
+        boxcar_lengths, return_max=True, normalization_func=lambda x: x
+    )
+    noises = calculate.calculate_noises_multi(
+        time_series, boxcar_array=boxcar_array, max_boxcar=max_boxcar
+    )
+    np.testing.assert_allclose(noises, 1 / np.sqrt(boxcar_lengths), rtol=0.3)
+
+
+def test_NoiseInfoResult():
+    """
+    Test the NoiseInfoResult dataclass
+    """
+    num_samps = 10
+    num_boxcars = 8
+    noises = rng.integers(10, size=num_samps * num_boxcars).reshape(
+        num_samps, num_boxcars
+    )
+    noise_info = calculate.NoiseInfoResult(
+        noises, np.array([2**x for x in range(0, num_boxcars)]), 4096
+    )
+    np.testing.assert_allclose(noise_info.mean_noise_levels, noises.mean(axis=0))
+    np.testing.assert_allclose(
[…]
