+++ will/calculate.py
@@ -5,10 +5,16 @@ Math functions.
 import logging
 import operator
 import warnings
-from typing import Callable
+from dataclasses import dataclass
+from functools import partial
+from typing import Callable, Tuple, Union
 
+import matplotlib.pyplot as plt
 import numpy as np
-from scipy import optimize, signal
+from jess.dispersion import dedisperse, delay_lost
+from rich.progress import track
+from scipy import interpolate, optimize, signal, stats
+from your import Your
 
 
 # pylint: disable=invalid-name
@@ -196,7 +202,7 @@ def generate_boxcar_array(
     boxcar_lengths: np.ndarray,
     normalization_func: Callable = np.sqrt,
     return_max: bool = False,
-) -> np.ndarray:
+) -> Union[np.ndarray, Tuple[np.ndarray, int]]:
     """
     Make a 2D array of boxcars for the given boxcar_lengths.
 
@@ -281,3 +287,223 @@ def boxcar_convolved(time_profile: np.ndarray, boxcar_widths: np.ndarray) -> np.
             convolved_profile = time_profile
         powers[j] = convolved_profile.max()
     return powers
+
+
+def median_line_detrend(array: np.ndarray, num_sample: int) -> np.ndarray:
+    """
+    Detrend an array by finding medians of sections and fitting a line
+    to these medians, then subtracting this line.
+
+    Args:
+        array - Array to detrend.
+
+        num_samples - Number of samples per median block.
+
+    Return:
+        array with trend subtracted.
+
+    Inspired by 'Advanced Architectures for Astrophysical Supercomputing',
+    section 4.2.3
+    """
+    len_array = len(array)
+    num_medians = np.ceil(len_array / num_sample).astype(int)
+    desired_length = num_medians * num_sample
+    to_length = desired_length - len_array
+
+    if to_length > 0:
+        half_length = to_length / 2
+        correct_array = np.pad(
+            array,
+            (np.floor(half_length).astype(int), np.ceil(half_length).astype(int)),
+            mode="reflect",
+        )
+    else:
+        correct_array = array
+
+    reshaped_array = correct_array.reshape(num_medians, -1)
+    medians = np.median(reshaped_array, axis=1)
+
+    offset = np.around(len_array / (2 * num_medians))
+    middles = np.linspace(offset, len_array - offset, num_medians, dtype=int)
+    interp_func = interpolate.interp1d(
+        middles, medians, fill_value="extrapolate", kind="linear"
+    )
+
+    full_samples = np.linspace(0, len_array, len_array, dtype=int)
+    return array - interp_func(full_samples)
+
+
+def calculate_noises_multi(
+    time_series: np.ndarray,
+    boxcar_array: np.ndarray,
+    max_boxcar: int,
+    smoothing_factor: int = 4,
+    scale_func: Callable = partial(stats.median_abs_deviation, scale="normal"),
+) -> np.ndarray:
+    """
+    Calculate the noise level of a time series over a set
+    of boxcars. Detrend the time series using median_line_detrend.
+    Convolve the detrened time series with the boxcar array,
+    Then use `scale_func` to calculate the noise level.
+
+    Args:
+        time_series - The dedispersed time series.
+
+        box_car_length - Length of the boxcars.
+
+        sigma - Return pulses with significance above
+                this
+
+        smoothing_factor - Detrend blocks are
+                           smoothing_factor*box_car_length long.
+
+        scale_func - The function used to calculate the measure of scale.
+                     must accept `axis=0` argument.
+
+    Returns:
+        Measures of scale for each boxcar in `boxcar array`.
+    """
+
+    flattened_times_series = median_line_detrend(
[…]
