### 2022-11-11 eeaf1c8
Added noise_info to README
